Refactored all modals.
