+++ client/components/core/forms/TeamsForm.vue
@@ -1,0 +1,74 @@
+<template>
+  <BaseForm
+    id="teams-form"
+    ref="baseForm"
+    button-label="Get a team of ultra fast coders"
+    :use-description="true"
+    @submit="handleSubmit"
+  >
+    <RadioList
+      id="teams"
+      label="Expected team size"
+      :options="options"
+      :required="true"
+      @select="handleSelectTeamSize"
+    />
+  </BaseForm>
+</template>
+
+<script>
+import BaseForm from '@/components/core/forms/BaseForm'
+import RadioList from '@/components/shared/UIRadioList'
+import sendEmailMixin from '@/mixins/sendEmailMixin'
+import exceptKeys from '@/helpers/exceptKeys'
+
+export default {
+  name: 'TeamsForm',
+  components: {
+    BaseForm,
+    RadioList,
+  },
+
+  mixins: [sendEmailMixin(304637, 'Teams')],
+
+  data() {
+    return {
+      selectedTeamSize: null,
+      options: [
+        {
+          value: 'less-five',
+          label: 'Less than 5',
+        },
+        {
+          value: 'from-five-to-ten',
+          label: 'From 5 to 10',
+        },
+        {
+          value: 'more-than-ten',
+          label: 'More than 10',
+        },
+      ],
+    }
+  },
+
+  methods: {
+    handleSelectTeamSize(teamSize) {
+      this.selectedTeamSize = teamSize
+    },
+
+    handleSubmit(formData) {
+      // from mixin
+      this.submitEmail({
+        ...exceptKeys(formData, ['description']),
+        selectedTeamSize: this.selectedTeamSize ? this.selectedTeamSize.label || '' : '',
+        projectDescription: formData.description,
+      })
+    },
+
+    reset() {
+      this.$refs.baseForm.reset()
+      this.selectedTeamSize = null
+    },
+  },
+}
+</script>
