+++ client/helpers/delay.js
@@ -1,0 +1,1 @@
+export default ms => new Promise(resolve => setTimeout(resolve, ms))
